demo(android): unify text + image + video + billboard into two-d-in-three-d (#2239 Batch 1)
Consolidates the retired `text`, `image`, `video`, and `billboard` demos
into a single `two-d-in-three-d` entry with a top segmented-button toggle:

- **Text** (default) — three stacked `TextNode` labels with editable
  content + font-size slider.
- **Image** — three-photo `ImageNode` gallery (procedural sky/sun/hill
  bitmaps) with a unison scale slider.
- **Video** — `VideoNode` streaming the bundled Big Buck Bunny MP4 with
  surface variants (Plane / Chrome cube / Reflective floor), cinematic
  vs orbit camera, and play/mute controls.
- **Billboard** — `BillboardNode` (faces camera) vs fixed `ImageNode`,
  planted on a checkered ground plane to make orbit-induced edge-on
  visible.

Each sub-mode keeps its own `SceneView` — the four scenes have different
camera framings, environments (Video needs HDR IBL + black skybox),
and pipelines. Inlined as private composables in the unified file.

### Deep-link aliases

`text`, `image`, `video`, `billboard` redirect to `two-d-in-three-d`
through `DeepLinkRouter.DEMO_ID_ALIASES`. The unified demo opens on its
default Text tab (tab pre-selection from alias is a planned follow-up).

`.maestro/android/content.yaml` keeps referring to the retired ids and
passes via the alias (only asserts the per-demo crash gate).

### Tests

- `DemoInteractionTest.textLabels_fontSizeSlider` — Text is the default
  tab, no extra tab tap.
- `DemoInteractionTest.image_scaleSlider` — adds `tap("Image")`.
- `DemoInteractionTest.video_initialRender` — adds `tap("Video")`.
- `DemoInteractionTest.billboard_visibilityChips` — adds `tap("Billboard")`.
- `DemoRenderingScreenshotTest.twoDInThreeDDemo_default_state` — unified
  slug. Stale `text_default.png`, `image_default.png`,
  `billboard_default.png` goldens removed; harness will re-capture.

### Quality gates

- ✅ `./gradlew :samples:android-demo:assembleDebug`
- ✅ `./gradlew :sceneview:detekt :arsceneview:detekt :sceneview-core:detekt`
- ✅ `bash samples/android-demo/scripts/collate-demos.sh` (58 fragments)

### Out of scope

- iOS mirror — Android-only per Batch 1 plan.
- Tab pre-selection from alias — follow-up.

Refs #2239 Batch 1.
