+++ changelog.d/2239-batch-1-two-d-in-three-d.md
@@ -1,0 +1,2 @@
+<!-- category: Changed -->
+- **Samples catalog: unified 2D in 3D demo ([#2239](https://github.com/sceneview/sceneview/issues/2239)).** The retired `text`, `image`, `video`, and `billboard` demos consolidated into a single `two-d-in-three-d` entry with a segmented-button toggle between Text (`TextNode` labels), Image (`ImageNode` photo gallery), Video (`VideoNode` streaming MP4 with surface variants + cinematic camera), and Billboard (`BillboardNode` vs fixed `ImageNode`) modes. Existing `sceneview://demo/text|image|video|billboard` deep links keep working via `DEMO_ID_ALIASES`.
