--- samples/android-demo/src/main/res/values/strings_demo_billboard.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `billboard` demo (registered by
-     `BillboardFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_billboard_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_billboard_title">Billboard</string>
-    <string name="demo_billboard_subtitle">Labels that face the camera</string>
-</resources>
