--- samples/android-demo/src/main/res/values/strings_demo_image.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `image` demo (registered by
-     `ImageFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_image_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_image_title">Image Planes</string>
-    <string name="demo_image_subtitle">Image planes in 3D space</string>
-    <string name="demo_image_detection">Image Detection</string>
-    <string name="demo_image_detection_desc">Real-world image recognition with AR overlay</string>
-</resources>
