--- samples/android-demo/src/main/res/values/strings_demo_video.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `video` demo (registered by
-     `VideoFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_video_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_video_title">Video Texture</string>
-    <string name="demo_video_subtitle">Video playback on a 3D surface</string>
-</resources>
